Coverity: fix SIGN_EXTENTION with clampBits
CID 1402217

Improve clampBits interface by fitting it more to the problem at hand:
This function is solely used to clamp some int into an unsigned short
(data type of 16 bit raw image). So have it return a ushort16 and add an
assert(n <= 16). This would effectively allow find a bug if someone
would accidentally clip to 17 bits and then assign the resulting 32-bit
int to a 16bit one.

+++ src/librawspeed/decoders/SrwDecoder.cpp
@@ -414,7 +414,7 @@ void SrwDecoder::decodeCompressed3(const TiffIFD* raw, int bits)
           value = &img[((i&0x7)<<1)+(i>>3)];
 
         diff = diff * (scale*2+1) + scale;
-        *value = clampBits((int64)(*value) + diff, bits);
+        *value = clampBits((int)*value + diff, bits);
       }
 
       img += 16;
